extract download label/path helpers to shared module
The two FilesToDownload twin components had identical 18-line copies of
getPathSuffix + downloadLabelByExt + getDownloadLabel. Lift them into
components/other/downloadLabel.ts so a future extract format adds one
entry in one place, not two.
