+++ src/components/other/FilesToDownload.tsx
@@ -1,19 +1,11 @@
 import styled from "styled-components";
-import { buttonsTitles } from "../../utils/texts";
 import Icon from "./Icons";
+import { getDownloadLabel } from "./downloadLabel";
 
 export interface FilesToDownloadProps {
   url: string;
   showFileName?: boolean;
 }
-
-const getPathSuffix = (raw: string) => {
-  try {
-    return new URL(raw).pathname.toLowerCase();
-  } catch (_) {
-    return raw.toLowerCase().split("?")[0];
-  }
-};
 
 const isExternalUrl = (raw: string) => {
   try {
@@ -22,21 +14,6 @@ const isExternalUrl = (raw: string) => {
   } catch (_) {
     return false;
   }
-};
-
-// Generated extracts arrive as .pdf, .zip (Geodatabase), or .geojson —
-// label the button accordingly so users know what they're downloading
-// before clicking. Anything else falls back to a generic "Atsisiųsti".
-const downloadLabelByExt: Record<string, string> = {
-  ".pdf": buttonsTitles.downloadPdf,
-  ".zip": buttonsTitles.downloadGdb,
-  ".geojson": buttonsTitles.downloadGeoJson,
-};
-
-const getDownloadLabel = (url: string) => {
-  const path = getPathSuffix(url);
-  const ext = Object.keys(downloadLabelByExt).find((e) => path.endsWith(e));
-  return ext ? downloadLabelByExt[ext] : buttonsTitles.download;
 };
 
 const FilesToDownload = ({ url, showFileName }: FilesToDownloadProps) => {
